+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - size and title of TabBtn
 - click handler of widgets
+- API endpoint to `/dtcd_workspaces`
 
 ## [0.6.0]